+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `esc` keybinding closes help and tag windows
 - Display a reason in the error message when the stream formats are not available for a video.
+- Don't generate instances file if there is no available Invidious instance.
 
 ### Changed
 - Lower the displayed precision of elapsed time on refresh completion.